Ensure that ping threads are closed when the ServerApp finishes.
Otherwise, the thread might keep running and evend send pings on a wrong
file descriptor. The reason is that file descriptors seem to be re-used
by the system. Now, if the ping thread sleeps during such a re-connect
and re-use of a file descriptor, it will simply keep pinging and
tickling on the new connection. This can easily lead to a lot of threads
that won't terminate.